requirements.txt
# Core dependencies for macOS Voice Bridge
whisper-cpp-python>=0.3.0
pyobjc-framework-CoreGraphics>=9.0
pyobjc-framework-Cocoa>=9.0
sounddevice>=0.4.6
pynput>=1.7.6
numpy>=1.24.0
